+++ imports/ui/components/LoginForm.jsx
@@ -25,7 +25,7 @@ export class LoginForm extends Tracker.Component {
       resetPasswordPath,
       profilePath,
       changePasswordPath
-    } = this.props;
+    } = props;
     // Set inital state.
     this.state = {
       message: null,
@@ -35,7 +35,9 @@ export class LoginForm extends Tracker.Component {
       onSignedInHook: props.onSignedInHook || Accounts.ui._options.onSignedInHook,
       onSignedOutHook: props.onSignedOutHook || Accounts.ui._options.onSignedOutHook,
       onPreSignUpHook: props.onPreSignUpHook || Accounts.ui._options.onPreSignUpHook,
-      onPostSignUpHook: props.onPostSignUpHook || Accounts.ui._options.onPostSignUpHook
+      onPostSignUpHook: props.onPostSignUpHook || Accounts.ui._options.onPostSignUpHook,
+      // Add default field values.
+      ...this.getDefaultFieldValues(),
     };
 
     // Listen for the user to login/logout.
@@ -53,12 +55,6 @@ export class LoginForm extends Tracker.Component {
     let changeState = Session.get(KEY_PREFIX + 'state');
     switch (changeState) {
       case 'enrollAccountToken':
-        this.setState({
-          formState: STATES.ENROLL_ACCOUNT
-        });
-        Session.set(KEY_PREFIX + 'state', null);
-        break;
-
       case 'resetPasswordToken':
         this.setState({
           formState: STATES.PASSWORD_CHANGE
@@ -75,10 +71,19 @@ export class LoginForm extends Tracker.Component {
     }
   }
 
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.formState != this.state.formState) {
+  componentWillReceiveProps(nextProps, nextContext) {
+    if (nextProps.formState && nextProps.formState !== this.state.formState) {
       this.setState({
-        formState: nextProps.formState
+        formState: nextProps.formState,
+        ...this.getDefaultFieldValues(),
+      });
+    }
+  }
+
+  componentDidUpdate(prevProps, prevState) {
+    if (!prevState.user !== !this.state.user) {
+      this.setState({
+        formState: this.state.user ? STATES.PROFILE : STATES.SIGN_IN
       });
     }
   }
@@ -171,17 +176,6 @@ export class LoginForm extends Tracker.Component {
     };
   }
 
-  getSetPasswordField() {
-    return {
-      id: 'newPassword',
-      hint: T9n.get('enterPassword'),
-      label: T9n.get('choosePassword'),
-      type: 'password',
-      required: true,
-      onChange: this.handleChange.bind(this, 'newPassword')
-    };
-  }
-
   handleChange(field, evt) {
     let value = evt.target.value;
     switch (field) {
@@ -191,6 +185,7 @@ export class LoginForm extends Tracker.Component {
         break;
     }
     this.setState({ [field]: value });
+    this.setDefaultFieldValues({ [field]: value });
   }
 
   fields() {
@@ -268,21 +263,24 @@ export class LoginForm extends Tracker.Component {
     }
 
     if (this.showPasswordChangeForm()) {
-      if (Meteor.isClient && !Accounts._loginButtonsSession.get('resetPasswordToken')) {
+      if (Meteor.isClient && !Accounts._loginButtonsSession.get('resetPasswordToken')
+        && !Accounts._loginButtonsSession.get('enrollAccountToken')) {
         loginFields.push(this.getPasswordField());
       }
       loginFields.push(this.getNewPasswordField());
     }
 
-    if (this.showEnrollAccountForm()) {
-      loginFields.push(this.getSetPasswordField());
-    }
-
-
     return _.indexBy(loginFields, 'id');
   }
 
   buttons() {
+    const {
+      loginPath = Accounts.ui._options.loginPath,
+      signUpPath = Accounts.ui._options.signUpPath,
+      resetPasswordPath = Accounts.ui._options.resetPasswordPath,
+      changePasswordPath = Accounts.ui._options.changePasswordPath,
+      profilePath = Accounts.ui._options.profilePath,
+    } = this.props;
     const { formState, waiting, user } = this.state;
     let loginButtons = [];
 
@@ -300,7 +298,7 @@ export class LoginForm extends Tracker.Component {
         id: 'switchToSignUp',
         label: T9n.get('signUp'),
         type: 'link',
-        href: this.props.signUpPath || Accounts.ui._options.signUpPath,
+        href: signUpPath,
         onClick: this.switchToSignUp.bind(this)
       });
     }
@@ -310,7 +308,7 @@ export class LoginForm extends Tracker.Component {
         id: 'switchToSignIn',
         label: T9n.get('signIn'),
         type: 'link',
-        href: this.props.loginPath || Accounts.ui._options.loginPath,
+        href: loginPath,
         onClick: this.switchToSignIn.bind(this)
       });
     }
@@ -320,7 +318,7 @@ export class LoginForm extends Tracker.Component {
         id: 'switchToPasswordReset',
         label: T9n.get('forgotPassword'),
         type: 'link',
-        href: this.props.resetPasswordPath || Accounts.ui._options.resetPasswordPath,
+        href: resetPasswordPath,
         onClick: this.switchToPasswordReset.bind(this)
       });
     }
@@ -335,7 +333,7 @@ export class LoginForm extends Tracker.Component {
         id: 'switchToChangePassword',
         label: T9n.get('changePassword'),
         type: 'link',
-        href: this.props.changePasswordPath || Accounts.ui._options.changePasswordPath,
+        href: changePasswordPath,
         onClick: this.switchToChangePassword.bind(this)
       });
     }
@@ -372,10 +370,10 @@ export class LoginForm extends Tracker.Component {
       });
     }
 
-    if (this.showPasswordChangeForm() || this.showEnrollAccountForm()) {
+    if (this.showPasswordChangeForm()) {
       loginButtons.push({
         id: 'changePassword',
-        label: (this.showPasswordChangeForm() ? T9n.get('changePassword') : T9n.get('setPassword')),
+        label: T9n.get('changePassword'),
         type: 'submit',
         disabled: waiting,
         onClick: this.passwordChange.bind(this)
@@ -385,7 +383,7 @@ export class LoginForm extends Tracker.Component {
         id: 'switchToSignOut',
         label: T9n.get('cancel'),
         type: 'link',
-        href: this.props.profilePath || Accounts.ui._options.profilePath,
+        href: profilePath,
         onClick: this.switchToSignOut.bind(this)
       });
     }
@@ -412,11 +410,6 @@ export class LoginForm extends Tracker.Component {
       && this.state.formState == STATES.PASSWORD_CHANGE);
   }
 
-  showEnrollAccountForm() {
-    return(Package['accounts-password']
-      && this.state.formState == STATES.ENROLL_ACCOUNT);
-  }
-
   showCreateAccountLink() {
     return this.state.formState == STATES.SIGN_IN && !Accounts._options.forbidClientAccountCreation && Package['accounts-password'];
   }
@@ -429,39 +422,91 @@ export class LoginForm extends Tracker.Component {
         passwordSignupFields());
   }
 
+  /**
+   * Helper to store field values while using the form.
+   */
+  setDefaultFieldValues(defaults) {
+    if (typeof defaults !== 'object') {
+      throw new Error('Argument to setDefaultFieldValues is not of type object');
+    } else if (localStorage) {
+      localStorage.setItem('accounts_ui', JSON.stringify({
+        ...this.getDefaultFieldValues(),
+        ...defaults,
+      }));
+    }
+  }
+
+  /**
+   * Helper to get field values when switching states in the form.
+   */
+  getDefaultFieldValues() {
+    if (localStorage) {
+      return JSON.parse(localStorage.getItem('accounts_ui') || null);
+    }
+  }
+
+  /**
+   * Helper to clear field values when signing in, up or out.
+   */
+  clearDefaultFieldValues() {
+    if (localStorage) {
+      localStorage.removeItem('accounts_ui');
+    }
+  }
+
   switchToSignUp(event) {
     event.preventDefault();
     this.setState({
       formState: STATES.SIGN_UP,
       message: null,
-      email: null
+      ...this.getDefaultFieldValues(),
     });
   }
 
   switchToSignIn(event) {
     event.preventDefault();
-    this.setState({ formState: STATES.SIGN_IN, message: null });
+    this.setState({
+      formState: STATES.SIGN_IN,
+      message: null,
+      ...this.getDefaultFieldValues(),
+    });
   }
 
   switchToPasswordReset(event) {
     event.preventDefault();
-    this.setState({ formState: STATES.PASSWORD_RESET, message: null });
+    this.setState({
+      formState: STATES.PASSWORD_RESET,
+      message: null,
+      ...this.getDefaultFieldValues(),
+    });
   }
 
   switchToChangePassword(event) {
     event.preventDefault();
-    this.setState({ formState: STATES.PASSWORD_CHANGE, message: null });
+    this.setState({
+      formState: STATES.PASSWORD_CHANGE,
+      message: null,
+      ...this.getDefaultFieldValues(),
+    });
   }
 
   switchToSignOut(event) {
     event.preventDefault();
-    this.setState({ formState: STATES.PROFILE, message: null });
+    this.setState({
+      formState: STATES.PROFILE,
+      message: null,
+    });
   }
 
   signOut() {
     Meteor.logout(() => {
-      this.setState({ formState: STATES.SIGN_IN, message: null, password: null });
+      this.setState({
+        formState: STATES.SIGN_IN,
+        message: null,
+        password: null,
+      });
       this.state.onSignedOutHook();
+      this.clearDefaultFieldValues();
     });
   }
 
@@ -470,7 +515,9 @@ export class LoginForm extends Tracker.Component {
       username = null,
       email = null,
       usernameOrEmail = null,
-      password
+      password,
+      formState,
+      onSubmitHook
     } = this.state;
 
     let loginSelector;
@@ -512,12 +559,14 @@ export class LoginForm extends Tracker.Component {
     }
 
     Meteor.loginWithPassword(loginSelector, password, (error, result) => {
+      onSubmitHook(error,formState);
       if (error) {
         this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"), 'error');
       }
       else {
         loginResultCallback(() => this.state.onSignedInHook());
         this.setState({ formState: STATES.PROFILE, message: null, password: null });
+        this.clearDefaultFieldValues();
       }
     });
   }
@@ -543,7 +592,7 @@ export class LoginForm extends Tracker.Component {
   }
 
   oauthSignIn(serviceName) {
-    const { formState, waiting, user } = this.state;
+    const { formState, waiting, user, onSubmitHook } = this.state;
     //Thanks Josh Owens for this one.
     function capitalService() {
       return serviceName.charAt(0).toUpperCase() + serviceName.slice(1);
@@ -564,10 +613,12 @@ export class LoginForm extends Tracker.Component {
       options.forceApprovalPrompt = Accounts.ui._options.forceApprovalPrompt[serviceName];
 
     loginWithService(options, (error) => {
+      onSubmitHook(error,formState);
       if (error) {
         this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"));
       } else {
         this.setState({ formState: STATES.PROFILE, message: '' });
+        this.clearDefaultFieldValues();
         loginResultCallback(() => {
           Meteor.setTimeout(() => this.state.onSignedInHook(), 100);
         });
@@ -582,7 +633,8 @@ export class LoginForm extends Tracker.Component {
       email = null,
       usernameOrEmail = null,
       password,
-      formState
+      formState,
+      onSubmitHook
     } = this.state;
 
     if (username !== null) {
@@ -620,7 +672,7 @@ export class LoginForm extends Tracker.Component {
     }
     else if (!validatePassword(password)) {
       this.showMessage(T9n.get("error.minChar").replace(/7/, Accounts.ui._options.minimumPasswordLength), 'warning');
-      this.state.onSubmitHook("error.minChar", formState);
+      onSubmitHook("error.minChar", formState);
       return;
     }
     else {
@@ -634,13 +686,14 @@ export class LoginForm extends Tracker.Component {
         if (error) {
           this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"), 'error');
           if (T9n.get(`error.accounts.${error.reason}`)) {
-            this.state.onSubmitHook(`error.accounts.${error.reason}`, formState);
+            onSubmitHook(`error.accounts.${error.reason}`, formState);
           }
           else {
-            this.state.onSubmitHook("Unknown error", formState);
+            onSubmitHook("Unknown error", formState);
           }
         }
         else {
+          onSubmitHook(null, formState);
           this.setState({
             formState: STATES.PROFILE,
             message: null,
@@ -648,6 +701,7 @@ export class LoginForm extends Tracker.Component {
           });
           let user = Accounts.user();
           loginResultCallback(this.state.onPostSignUpHook.bind(this, _options, user));
+          this.clearDefaultFieldValues();
         }
 
         this.setState({ waiting: false });
@@ -668,7 +722,9 @@ export class LoginForm extends Tracker.Component {
     const {
       email = '',
       usernameOrEmail = '',
-      waiting
+      waiting,
+      formState,
+      onSubmitHook
     } = this.state;
 
     if (waiting) {
@@ -684,8 +740,9 @@ export class LoginForm extends Tracker.Component {
         }
         else {
           this.showMessage(T9n.get("info.emailSent"), 'success', 5000);
+          this.clearDefaultFieldValues();
         }
-
+        onSubmitHook(error, formState);
         this.setState({ waiting: false });
       });
     }
@@ -698,25 +755,31 @@ export class LoginForm extends Tracker.Component {
         }
         else {
           this.showMessage(T9n.get("info.emailSent"), 'success', 5000);
+          this.clearDefaultFieldValues();
         }
-
+        onSubmitHook(error, formState);
         this.setState({ waiting: false });
       });
     }
     else {
+      let errMsg = null;
       if (_.contains([ "USERNAME_AND_EMAIL_NO_PASSWORD" ], passwordSignupFields())) {
-        this.showMessage(T9n.get("error.accounts.Invalid email or username"), 'warning');
+        errMsg = T9n.get("error.accounts.Invalid email or username");
       }
       else {
-        this.showMessage(T9n.get("error.accounts.Invalid email"), 'warning');
+        errMsg = T9n.get("error.accounts.Invalid email");
       }
+      this.showMessage(errMsg,'warning');
+      onSubmitHook(errMsg, formState);
     }
   }
 
   passwordReset() {
     const {
       email = '',
-      waiting
+      waiting,
+      formState,
+      onSubmitHook
     } = this.state;
 
     if (waiting) {
@@ -732,24 +795,31 @@ export class LoginForm extends Tracker.Component {
         }
         else {
           this.showMessage(T9n.get("info.emailSent"), 'success', 5000);
+          this.clearDefaultFieldValues();
         }
-
+        onSubmitHook(error, formState);
         this.setState({ waiting: false });
       });
     }
     else {
-      this.showMessage(T9n.get("error.accounts.Invalid email"), 'warning');
+      const errMsg = T9n.get("error.accounts.Invalid email");
+      onSubmitHook(errMsg, formState);
+      this.showMessage(errMsg,'warning');
     }
   }
 
   passwordChange() {
     const {
       password,
-      newPassword
+      newPassword,
+      formState,
+      onSubmitHook
     } = this.state;
 
     if ( !validatePassword(newPassword) ){
-      this.showMessage(T9n.get("error.minChar").replace(/7/, Accounts.ui._options.minimumPasswordLength), 'warning');
+      const errMsg = T9n.get("error.minChar").replace(/7/, Accounts.ui._options.minimumPasswordLength);
+      this.showMessage(errMsg,'warning');
+      onSubmitHook(errMsg,formState);
       return;
     }
 
@@ -761,9 +831,11 @@ export class LoginForm extends Tracker.Component {
       Accounts.resetPassword(token, newPassword, (error) => {
         if (error) {
           this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"), 'error');
+          onSubmitHook(error, formState);
         }
         else {
           this.showMessage(T9n.get('info.passwordChanged'), 'success', 5000);
+          onSubmitHook(null, formState);
           this.setState({ formState: STATES.PROFILE });
           Accounts._loginButtonsSession.set('resetPasswordToken', null);
           Accounts._loginButtonsSession.set('enrollAccountToken', null);
@@ -774,10 +846,13 @@ export class LoginForm extends Tracker.Component {
       Accounts.changePassword(password, newPassword, (error) => {
         if (error) {
           this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"), 'error');
+          onSubmitHook(error, formState);
         }
         else {
           this.showMessage(T9n.get('info.passwordChanged'), 'success', 5000);
+          onSubmitHook(null, formState);
           this.setState({ formState: STATES.PROFILE });
+          this.clearDefaultFieldValues();
         }
       });
     }
@@ -796,20 +871,16 @@ export class LoginForm extends Tracker.Component {
     }
   }
 
-  componentDidUpdate(prevProps, prevState) {
-    if (!prevState.user !== !this.state.user) {
-      this.setState({
-        formState: this.state.user ? STATES.PROFILE : STATES.SIGN_IN
-      });
-    }
-  }
-
   render() {
     this.oauthButtons();
-    return <Accounts.ui.Form oauthServices={this.oauthButtons()}
-                             fields={this.fields()} 
-                             buttons={this.buttons()}
-                             {...this.state} />;
+    return (
+      <Accounts.ui.Form
+        oauthServices={this.oauthButtons()}
+        fields={this.fields()} 
+        buttons={this.buttons()}
+        {...this.state}
+      />
+    );
   }
 };