+++ package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: 'std:accounts-ui',
-  version: '1.2.6',
-  summary: 'Accounts UI for React in Meteor 1.3',
+  version: '1.2.11',
+  summary: 'Accounts UI for React in Meteor 1.3+',
   git: 'https://github.com/studiointeract/accounts-ui',
   documentation: 'README.md'
 });